requirements.txt
Django>=4.2
djangorestframework
sorl-thumbnail
django-taggit
channels
channels-redis
psycopg2-binary
django-humanize
python-decouple
